feat(TDI-39100): Move CouchbaseSourceOrSink to runtime package
